Maybe fix tests, bring mock back into WMFDataTests
